Correct rating algorithm. Checks string id and calculates double instead of integer value

+++ MovieService/src/main/java/rs/uns/acs/ftn/models/Movie.java
@@ -231,7 +231,7 @@ public class Movie implements Serializable {
 			counter+=1;
 			sum+=rat.getValue();
 		}
-		double ratingAvg = sum/counter;
+		double ratingAvg = ((double)sum)/counter;
 		return ratingAvg;
 	}
